Move index page into its own file
